Rely on envvar only to determine prod mode
The `golem` helpers rely on the `"golem.app.prod"` option, but not the
envvar `GOLEM_CONFIG_ACTIVE`, which is what we're using. We were
synchronising this setting before, but as we're not relying on the
option, it's probably better to just avoid any possible confusion by not
relying on the `golem` helpers.